# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/compose/episodes/{id}/compose-status": {
      "get": {
        "operationId": "ComposeController_getEpisodeComposeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "compose"
        ]
      }
    },
    "/api/v1/compose/storyboards/{id}/compose": {
      "post": {
        "operationId": "ComposeController_composeStoryboard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "compose"
        ]
      }
    },
    "/api/v1/compose/episodes/{id}/compose-all": {
      "post": {
        "operationId": "ComposeController_composeEpisode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "compose"
        ]
      }
    },
...
    "/api/v1/characters/{id}": {
      "put": {
        "operationId": "CharactersController_updateCharacter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "characters"
        ]
      },
      "delete": {
        "operationId": "CharactersController_deleteCharacter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/v1/canvases/{id}": {
      "get": {
        "operationId": "CanvasController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "canvas"
        ]
      },
      "patch": {
        "operationId": "CanvasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "canvas"
        ]
      },
      "delete": {
        "operationId": "CanvasController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "canvas"
        ]
      }
    },
    "/api/v1/canvases/{id}/duplicate": {
      "post": {
        "operationId": "CanvasController_duplicate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "canvas"
        ]
      }
    },
    "/api/v1/canvases/{id}/save": {
      "post": {
        "operationId": "CanvasController_save",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "canvas"
        ]
      }
    },
...
  },
  "info": {
    "title": "XIAOCHUANG Backend API",
    "description": "V2.0 unified backend for web and admin clients",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}